requirements.txt
requests 
google_images_download
asyncio 
pip
discord.py
bs4